pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-http-python-test"
description = "Source implementation for http-python-test."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/http-python-test"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_http_python_test", from="src"}, {include = "main.py", from = "src"} ]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^0"

[tool.poetry.scripts]
source-http-python-test = "source_http_python_test.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "^1.9.3"
pytest-mock = "^3.6.1"
pytest = "^6.1"
